app/requirements.txt
arrow==1.1.1
asgiref==3.4.1
autopep8==1.5.7
binaryornot==0.4.4
boto3==1.18.4
botocore==1.21.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
cookiecutter==1.7.3
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-storages==1.11.1
gunicorn==20.1.0
idna==3.2
Jinja2==3.0.1
jinja2-time==0.2.0
jmespath==0.10.0
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==9.2.0
poyo==0.5.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
